Example cleanup
