+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 This workspace publishes multiple crates. Each crate maintains its own changelog:
 
+- [`linktime/CHANGELOG.md`](linktime/CHANGELOG.md)
 - [`ctor/CHANGELOG.md`](ctor/CHANGELOG.md)
 - [`dtor/CHANGELOG.md`](dtor/CHANGELOG.md)
 - [`link-section/CHANGELOG.md`](link-section/CHANGELOG.md)
