+++ ctor/Cargo.toml
@@ -10,6 +10,8 @@ readme = "README.md"
 categories = ["no-std"]
 rust-version = "1.60"
 
+exclude = ["tests"]
+
 [features]
 default = ["std", "proc_macro", "priority"]
