+++ ctor/README.md
@@ -10,7 +10,7 @@ The crate is part of the [`linktime`](https://crates.io/crates/linktime) project
 | `link-section` | [![docs.rs](https://docs.rs/link-section/badge.svg)](https://docs.rs/link-section) | [![crates.io](https://img.shields.io/crates/v/link-section.svg)](https://crates.io/crates/link-section) |
 # ctor
 Module initialization functions for Rust (like `__attribute__((constructor))` in
-C/C++) for Linux, OSX, Windows, WASM, BSD-likes, and many others.
+C/C++) for Linux, macOS, Windows, WASM, BSD-likes, and many others.
 
 ```rust
 use ctor::ctor;
@@ -36,7 +36,7 @@ should have minimal effect on compilation time.
 
 ## Support
 
-This library works and is regularly tested on Linux, OSX, Windows, and FreeBSD,
+This library works and is regularly tested on Linux, macOS, Windows, and FreeBSD,
 with both `+crt-static` and `-crt-static` and `bin`/`cdylib` outputs.
 
 Contributions to support other platforms or improve testing are welcome.
@@ -44,7 +44,7 @@ Contributions to support other platforms or improve testing are welcome.
 | OS           | Supported | CI Tested |
 | ------------ | --------- | --------- |
 | Linux        | ✅        | ✅        |
-| OSX          | ✅        | ✅        |
+| macOS        | ✅        | ✅        |
 | Windows      | ✅        | ✅        |
 | FreeBSD      | ✅        | ✅        |
 | WASM         | ✅        | ✅        |
@@ -53,7 +53,7 @@ Contributions to support other platforms or improve testing are welcome.
 | DragonFlyBSD | ✅        | -         |
 | Illumos      | ✅        | -         |
 | Android      | ✅        | -         |
-| iOS          | ✅        | -         |
+| iOS/tvOS/etc | ✅        | -         |
 | AIX          | ✅        | -         |
 | Haiku        | ✅        | -         |
 | VxWorks      | ✅        | -         |
